requirements.txt
Django==4.2.0
djangorestframework==3.14.0
python-decouple==3.8
celery==5.3.0
redis==4.5.4
python-telegram-bot==20.3 